ehance os.Root tests
 - add a test for a path traversal
 - use os.ReadFile for assertions
